Fix a bug in UK drainage calculations
If a soil_DT50 was specified, and the application date was later than
the beginning of the drainage period (1 October), PECsw would increase.
Now the amount available at the time of the drainage period is equal
to the applied amount (minus interception) if application is later than
the beginning of the drainage period. Fixes #4. Thanks to @lutzeli for
spotting the problem.